change task tag to 4bit; optimize bool and

+++ include/base/AbstractBatchExecutor.h
@@ -14,7 +14,7 @@ public:
     std::vector<int64_t> _yis{};
     std::vector<int64_t> _zis{};
 
-    AbstractBatchExecutor(int width, int16_t taskTag, int16_t msgTagOffset)
+    AbstractBatchExecutor(int width, int taskTag, int msgTagOffset)
         : AbstractSecureExecutor(width, taskTag, msgTagOffset) {
     }
 };
