+++ include/compute/batch/arith/ArithBatchExecutor.h
@@ -9,9 +9,9 @@
 
 class ArithBatchExecutor : public AbstractBatchExecutor {
 public:
-    ArithBatchExecutor(std::vector<int64_t>& zs, int width, int16_t taskTag, int16_t msgTagOffset, int clientRank);
+    ArithBatchExecutor(std::vector<int64_t>& zs, int width, int taskTag, int msgTagOffset, int clientRank);
 
-    ArithBatchExecutor(std::vector<int64_t>& xs, std::vector<int64_t>& ys, int width, int16_t taskTag, int16_t msgTagOffset, int clientRank);
+    ArithBatchExecutor(std::vector<int64_t>& xs, std::vector<int64_t>& ys, int width, int taskTag, int msgTagOffset, int clientRank);
 
     ArithBatchExecutor *reconstruct(int clientRank) override;
