+++ include/compute/batch/bool/BoolBatchExecutor.h
@@ -4,14 +4,14 @@
 
 #ifndef BOOLBATCHEXECUTOR_H
 #define BOOLBATCHEXECUTOR_H
-#include "base/AbstractBatchExecutor.h"
+#include "../../../base/AbstractBatchExecutor.h"
 
 
 class BoolBatchExecutor : public AbstractBatchExecutor {
 public:
-    BoolBatchExecutor(std::vector<int64_t>& zs, int width, int16_t taskTag, int16_t msgTagOffset, int clientRank);
+    BoolBatchExecutor(std::vector<int64_t>& zs, int width, int taskTag, int msgTagOffset, int clientRank);
 
-    BoolBatchExecutor(std::vector<int64_t>& xs, std::vector<int64_t>& ys, int width, int16_t taskTag, int16_t msgTagOffset, int clientRank);
+    BoolBatchExecutor(std::vector<int64_t>& xs, std::vector<int64_t>& ys, int width, int taskTag, int msgTagOffset, int clientRank);
 
     BoolBatchExecutor *reconstruct(int clientRank) override;
