+++ include/compute/single/bool/BoolAndExecutor.h
@@ -16,13 +16,13 @@ public:
     inline static std::atomic_int64_t _totalTime = 0;
 
 public:
-    BoolAndExecutor(int64_t x, int64_t y, int l, int16_t taskTag, int16_t msgTagOffset,
+    BoolAndExecutor(int64_t x, int64_t y, int l, int taskTag, int msgTagOffset,
                        int clientRank) : BoolExecutor(x, y, l, taskTag, msgTagOffset, clientRank) {
     }
 
     BoolAndExecutor *execute() override;
 
-    [[nodiscard]] static int16_t msgTagCount(int width);
+    [[nodiscard]] static int msgTagCount(int width);
 
     BoolAndExecutor *setBmt(BitwiseBmt *bmt);
 };
