+++ include/compute/single/bool/BoolEqualExecutor.h
@@ -14,7 +14,7 @@ private:
     BitwiseBmt *_bmt{};
 
 public:
-    BoolEqualExecutor(int64_t x, int64_t y, int l, int16_t taskTag, int16_t msgTagOffset,
+    BoolEqualExecutor(int64_t x, int64_t y, int l, int taskTag, int msgTagOffset,
                        int clientRank) : BoolExecutor(x, y, l, taskTag, msgTagOffset, clientRank) {}
 
     BoolEqualExecutor *execute() override;
