+++ include/compute/single/bool/BoolMutexExecutor.h
@@ -17,13 +17,13 @@ public:
     inline static std::atomic_int64_t _totalTime = 0;
 
 public:
-    BoolMutexExecutor(int64_t x, int64_t y, bool cond, int width, int16_t taskTag, int16_t msgTagOffset, int clientRank);
+    BoolMutexExecutor(int64_t x, int64_t y, bool cond, int width, int taskTag, int msgTagOffset, int clientRank);
 
     BoolMutexExecutor *execute() override;
 
     BoolMutexExecutor *setBmts(std::vector<BitwiseBmt> *bmts);
 
-    static int16_t msgTagCount(int l);
+    static int msgTagCount(int l);
 
     static int bmtCount(int width);
 };
