+++ include/compute/single/bool/BoolToArithExecutor.h
@@ -10,7 +10,7 @@
 class BoolToArithExecutor : public BoolExecutor {
 public:
     // Temporarily lend zi for xi preparation in super constructor.
-    BoolToArithExecutor(int64_t xi, int l, int16_t taskTag, int16_t msgTagOffset, int clientRank) : BoolExecutor(
+    BoolToArithExecutor(int64_t xi, int l, int taskTag, int msgTagOffset, int clientRank) : BoolExecutor(
         xi, l, taskTag, msgTagOffset, clientRank) {
         _xi = _zi;
         _zi = 0;
@@ -18,7 +18,7 @@ public:
 
     BoolToArithExecutor *execute() override;
 
-    [[nodiscard]] static int16_t msgTagCount(int width);
+    [[nodiscard]] static int msgTagCount(int width);
 
     BoolToArithExecutor *reconstruct(int clientRank) override;
 };
