+++ include/compute/single/bool/BoolXorExecutor.h
@@ -8,13 +8,13 @@
 
 class BoolXorExecutor : public BoolExecutor {
 public:
-    BoolXorExecutor(bool x, bool y, int l, int16_t taskTag, int16_t msgTagOffset,
+    BoolXorExecutor(bool x, bool y, int l, int taskTag, int msgTagOffset,
                        int clientRank) : BoolExecutor(x, y, l, taskTag, msgTagOffset, clientRank) {
     };
 
     BoolXorExecutor *execute() override;
 
-    [[nodiscard]] static int16_t msgTagCount();
+    [[nodiscard]] static int msgTagCount();
 };
