+++ include/ot/BaseOtExecutor.h
@@ -29,13 +29,13 @@ private:
 public:
     // _m0 and _m1 are for sender (invalid for receiver)
     // i is for receiver (invalid for sender)
-    explicit BaseOtExecutor(int sender, int64_t m0, int64_t m1, int choice, int l, int16_t taskTag, int16_t msgTagOffset);
+    explicit BaseOtExecutor(int sender, int64_t m0, int64_t m1, int choice, int l, int taskTag, int msgTagOffset);
 
-    explicit BaseOtExecutor(int bits, int sender, int64_t m0, int64_t m1, int choice, int l, int16_t taskTag, int16_t msgTagOffset);
+    explicit BaseOtExecutor(int bits, int sender, int64_t m0, int64_t m1, int choice, int l, int taskTag, int msgTagOffset);
 
     BaseOtExecutor *execute() override;
 
-    [[nodiscard]] static int16_t msgTagCount();
+    [[nodiscard]] static int msgTagCount();
 
 private:
     // methods for sender
