+++ include/ot/RandOtBatchExecutor.h
@@ -12,11 +12,11 @@ public:
     inline static std::atomic_int64_t _totalTime = 0;
 
 public:
-    RandOtBatchExecutor(int sender, std::vector<int64_t> *ms0, std::vector<int64_t> *ms1, std::vector<int> *choices, int l, int16_t taskTag, int16_t msgTagOffset) : AbstractOtBatchExecutor(sender, ms0, ms1, choices, l, taskTag, msgTagOffset) {}
+    RandOtBatchExecutor(int sender, std::vector<int64_t> *ms0, std::vector<int64_t> *ms1, std::vector<int> *choices, int l, int taskTag, int msgTagOffset) : AbstractOtBatchExecutor(sender, ms0, ms1, choices, l, taskTag, msgTagOffset) {}
 
     RandOtBatchExecutor *execute() override;
 
-    static int16_t msgTagCount();
+    static int msgTagCount();
 };
