+++ include/ot/base/AbstractOtBatchExecutor.h
@@ -17,7 +17,7 @@ protected:
 
 public:
     AbstractOtBatchExecutor(int sender, std::vector<int64_t> *ms0, std::vector<int64_t> *ms1, std::vector<int> *choices,
-                            int l, int16_t taskTag, int16_t msgTagOffset);
+                            int l, int taskTag, int msgTagOffset);
 
     AbstractOtBatchExecutor *reconstruct(int clientRank) override;
 };
